chore: add app_localizations
