+++ test/helpers/helpers.dart
@@ -1,16 +1,1 @@
-import 'package:flutter_test/flutter_test.dart';
-import 'package:hydrated_bloc/hydrated_bloc.dart';
-import 'package:mocktail/mocktail.dart';
-
-class MockStorage extends Mock implements Storage {}
-
-late Storage hydratedStorage;
-
-void initHydratedStorage() {
-  TestWidgetsFlutterBinding.ensureInitialized();
-  hydratedStorage = MockStorage();
-  when(
-    () => hydratedStorage.write(any(), any<dynamic>()),
-  ).thenAnswer((_) async {});
-  HydratedBloc.storage = hydratedStorage;
-}
+export 'pump_app.dart';
