+++ test/helpers/pump_app.dart
@@ -1,1 +1,30 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_test/flutter_test.dart';
+import 'package:hydrated_bloc/hydrated_bloc.dart';
+import 'package:mocktail/mocktail.dart';
+import 'package:my_coffees/l10n/l10n.dart';
 
+class MockStorage extends Mock implements Storage {}
+
+late Storage hydratedStorage;
+
+void initHydratedStorage() {
+  TestWidgetsFlutterBinding.ensureInitialized();
+  hydratedStorage = MockStorage();
+  when(
+    () => hydratedStorage.write(any(), any<dynamic>()),
+  ).thenAnswer((_) async {});
+  HydratedBloc.storage = hydratedStorage;
+}
+
+extension PumpApp on WidgetTester {
+  Future<void> pumpApp(Widget widget) {
+    return pumpWidget(
+      MaterialApp(
+        localizationsDelegates: AppLocalizations.localizationsDelegates,
+        supportedLocales: AppLocalizations.supportedLocales,
+        home: widget,
+      ),
+    );
+  }
+}
